requirements.txt
# runtime requirements

# Python Clarity Client API
-e git+https://github.com/crukci-bioinformatics/claritypy-client.git@v0.4.4#egg=claritypy-glsclient

# requests security package extra
requests[security]

# config
pyyaml
